Removed version.nsi.  It's no longer needed.  The version is now defined in xsupextras.nsi, and can be overridden with the /DVERSION=<whatever> flag.
Added open1xdebug.nsi to allow packaging of the various PDB files.  We'll need to figure out a more long-term way of handling this, but for the short term this provides an excellent low-cost way of ensuring that the PDB files are always in sync with their executable counterparts.

Wrapped thirdparty and debugfile directives in easy-to-disable defines.